+++ src/AddItem/AddItem.js
@@ -1,20 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import style from './AddItem.scss';
-import Tooltip from '../Tooltip';
 import AddMedia from 'wix-ui-icons-common/system/AddMedia';
-import WixComponent from '../BaseComponents/WixComponent';
 import classNames from 'classnames';
 
-const tooltipCommonProps = {
+import WixComponent from '../BaseComponents/WixComponent';
+import Tooltip from '../Tooltip';
+import style from './AddItem.scss';
+
+const DEFAULT_TOOLTIP_PROPS = {
   showDelay: 0,
-  theme: 'dark',
   hideDelay: 0,
+  theme: 'dark',
   align: 'center',
   placement: 'top'
 };
 
-const ratioClasses = {
+const RATIO_CLASSES = {
   '16/9': style.ratio16x9,
   '3/4': style.ratio3x4,
   '4/3': style.ratio4x3,
@@ -35,13 +36,19 @@ class AddItem extends WixComponent {
       tooltipContent,
       aspectRatio
     } = this.props;
-    const ratio = !height && ratioClasses[aspectRatio];
+    const ratio = !height && RATIO_CLASSES[aspectRatio];
+    const tooltipProps = {
+      ...DEFAULT_TOOLTIP_PROPS,
+      content: tooltipContent,
+      ...this.props.tooltipProps
+    };
+
     return (
       <div className={classNames(ratio, style.box)} style={{height}} >
         <div className={style.container} onClick={onClick} data-hook="add-container">
           {
-            tooltipContent ?
-              <Tooltip content={tooltipContent} dataHook="add-tooltip" {...tooltipCommonProps}>
+            tooltipProps.content ?
+              <Tooltip dataHook="add-tooltip" {...tooltipProps}>
                 {renderInnerAddItem()}
               </Tooltip> :
               renderInnerAddItem()
@@ -53,13 +60,15 @@ class AddItem extends WixComponent {
 }
 
 AddItem.propTypes = {
-  /** Funciton called upon click */
+  /** Function called upon click */
   onClick: PropTypes.func,
-  /** The elemnt's asspect ratio   */
-  aspectRatio: PropTypes.oneOf(Object.keys(ratioClasses)),
+  /** The element's aspect ratio */
+  aspectRatio: PropTypes.oneOf(Object.keys(RATIO_CLASSES)),
   /** Element's height - overrides the asspect ratio */
   height: PropTypes.number,
-  /** Content of the tooltip */
+  /** Tooltip props, leave undefined for no tooltip */
+  tooltipProps: PropTypes.shape(Tooltip.propTypes),
+  /** Content of the tooltip, leave undefined for no tooltip */
   tooltipContent: PropTypes.string
 };