+++ src/FormField/FormField.driver.js
@@ -1,43 +1,29 @@
-import {resolveIn} from '../../test/utils';
-import {tooltipTestkitFactory} from '../../testkit';
+import tooltipDriverFactory from '../Tooltip/Tooltip.driver';
 
-const findByHook = (element, hook) =>
-  element.querySelector(`[data-hook*="${hook}"]`);
+const formFieldDriver = ({wrapper, element}) => {
+  const byHook = hook => element.querySelector(`[data-hook*="${hook}"]`);
+  const charactersCounter = () => byHook('formfield-counter');
 
-const getInfoIcon = element =>
-  findByHook(element, 'formfield-infoicon');
-
-const getCharactersCounter = element =>
-  findByHook(element, 'formfield-counter');
-
-const formFieldDriver = ({element}) => ({
-  exists: () => !!element,
-  element: () => element,
-  getChildren: () => findByHook(element, 'formfield-children'),
-  getLabel: () => findByHook(element, 'formfield-label'),
-  isRequired: () => !!findByHook(element, 'formfield-asterisk'),
-  getLengthLeft: () => {
-    const counter = getCharactersCounter(element);
-    return counter ? parseInt(counter.innerHTML, 10) : null;
-  },
-  isLengthExceeded: () => {
-    const counter = getCharactersCounter(element);
-    if (counter) {
-      const length = parseInt(counter.innerHTML, 10);
-      return length < 0;
-    }
-    return false;
-  },
-  getInfoContent: () => {
-    const tooltipDriver = tooltipTestkitFactory({
-      wrapper: getInfoIcon(element),
-      dataHook: 'formfield-infotooltip'
-    });
-
-    tooltipDriver.mouseEnter();
-    return resolveIn(500)
-      .then(() => tooltipDriver.getContent());
-  }
-});
+  return {
+    exists: () => !!element,
+    element: () => element,
+    getChildren: () => byHook('formfield-children'),
+    getLabel: () => byHook('formfield-label'),
+    isRequired: () => !!byHook('formfield-asterisk'),
+    getLengthLeft: () => {
+      const counter = charactersCounter();
+      return counter ? parseInt(counter.innerHTML, 10) : null;
+    },
+    isLengthExceeded: () => {
+      const counter = charactersCounter();
+      if (counter) {
+        const length = parseInt(counter.innerHTML, 10);
+        return length < 0;
+      }
+      return false;
+    },
+    getInfoContent: () => tooltipDriverFactory({wrapper, element: byHook('formfield-infotooltip')}).hoverAndGetContent()
+  };
+};
 
 export default formFieldDriver;