+++ src/InputArea/InputArea.js
@@ -125,9 +125,9 @@ class InputArea extends WixComponent {
     this.textArea && this.textArea.select();
   }
 
-  _onFocus() {
+  _onFocus(e) {
     this.setState({focus: true});
-    this.props.onFocus && this.props.onFocus();
+    this.props.onFocus && this.props.onFocus(e);
 
     if (this.props.autoSelect) {
       // Set timeout is needed here since onFocus is called before react