+++ src/Layout/index.js
@@ -1,0 +1,4 @@
+import Layout from './Layout';
+import Cell from './Cell';
+
+export {Layout, Cell};