+++ src/Table/Toolbar/index.js
@@ -1,2 +1,4 @@
-export * from './Toolbar';
-export * from './TableToolbar';
+import deprecationLog from '../../utils/deprecationLog';
+
+deprecationLog('Importing TableToolbar from wix-style-react/dist/src/components/Table/Toolbar is deprecated. Please import from wix-style-react/TableToolbar');
+export * from '../../TableToolbar';