+++ stories/Layout/story-settings.js
@@ -1,0 +1,10 @@
+const stories = {
+  category: 'Components/Layout',
+  examplesCategory: 'Components/Layout/Examples',
+  holyGrailLayout: 'Holy Grail Layout',
+  listOfCards: 'List of Cards',
+  mainAndSide: 'Main and Side',
+  form: 'Form'
+};
+
+export default stories;